Move vizBuilderConfig under context folder
