+++ packages/admin-panel/src/VizBuilderApp/context/index.js
@@ -1,1 +1,2 @@
 export * from './PreviewData';
+export * from './VizBuilderConfig';
